pyproject.toml
[project]
name = "got-rag-chatbot"
version = "0.1.0"
description = "A chatbot leveraging RAG techniques for enhanced responses."
authors = [
    { name = "Diego Páez", email = "diripar8@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.6",
    "en-core-web-sm",
    "fastapi>=0.125.0",
    "ipykernel>=7.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-google-genai>=4.0.0",
    "langchain-huggingface>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "matplotlib>=3.10.7",
    "mwparserfromhell>=0.7.2",
    "neo4j>=6.0.3",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.0",
    "spacy>=3.8.11",
    "streamlit>=1.52.1",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
